refactor: subscribe 패키지 정리
