+++ src/main/java/maeilmail/subscribe/CodeGenerator.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.core;
+package maeilmail.subscribe;
 
 import java.util.Random;
 import org.springframework.stereotype.Component;
