+++ src/main/java/maeilmail/subscribe/SubscribeApi.java
@@ -1,9 +1,6 @@
-package maeilmail.subscribe.api;
+package maeilmail.subscribe;
 
 import lombok.RequiredArgsConstructor;
-import maeilmail.subscribe.core.SubscribeQuestionService;
-import maeilmail.subscribe.core.request.SubscribeQuestionRequest;
-import maeilmail.subscribe.core.request.VerifyEmailRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
