+++ src/main/java/maeilmail/subscribe/SubscribeQuestionRequest.java
@@ -1,0 +1,6 @@
+package maeilmail.subscribe;
+
+import java.util.List;
+
+record SubscribeQuestionRequest(String email, List<String> category, String code) {
+}
