+++ src/main/java/maeilmail/subscribe/SubscribeRepository.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.core;
+package maeilmail.subscribe;
 
 import java.time.LocalDateTime;
 import java.util.List;
