+++ src/main/java/maeilmail/subscribe/TemporalSubscribeRepository.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.core;
+package maeilmail.subscribe;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
