+++ src/main/java/maeilmail/subscribe/VerifyEmailRequest.java
@@ -1,0 +1,4 @@
+package maeilmail.subscribe;
+
+record VerifyEmailRequest(String email) {
+}
