--- src/main/java/maeilmail/subscribe/core/request/SubscribeQuestionRequest.java
@@ -1,6 +1,0 @@
-package maeilmail.subscribe.core.request;
-
-import java.util.List;
-
-public record SubscribeQuestionRequest(String email, List<String> category, String code) {
-}
