--- src/main/java/maeilmail/subscribe/core/request/VerifyEmailRequest.java
@@ -1,4 +1,0 @@
-package maeilmail.subscribe.core.request;
-
-public record VerifyEmailRequest(String email) {
-}
