+++ src/main/java/maeilmail/subscribequestion/policy/WeekdaysPersonalSequenceChoicePolicy.java
@@ -5,7 +5,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
-import maeilmail.subscribe.core.Subscribe;
+import maeilmail.subscribe.Subscribe;
 import maeilmail.subscribequestion.ChoiceQuestionPolicy;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
