+++ src/test/java/maeilmail/statistics/StatisticsServiceTest.java
@@ -8,8 +8,8 @@ import java.util.stream.IntStream;
 import maeilmail.question.Question;
 import maeilmail.question.QuestionCategory;
 import maeilmail.question.QuestionRepository;
-import maeilmail.subscribe.core.Subscribe;
-import maeilmail.subscribe.core.SubscribeRepository;
+import maeilmail.subscribe.Subscribe;
+import maeilmail.subscribe.SubscribeRepository;
 import maeilmail.subscribequestion.SubscribeQuestion;
 import maeilmail.subscribequestion.SubscribeQuestionRepository;
 import maeilmail.support.IntegrationTestSupport;
