+++ src/test/java/maeilmail/subscribe/SubscribeSequenceSchedulerTest.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.core;
+package maeilmail.subscribe;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
