+++ src/test/java/maeilmail/subscribequestion/policy/SequenceChoicePolicyTest.java
@@ -8,8 +8,8 @@ import maeilmail.question.Question;
 import maeilmail.question.QuestionCategory;
 import maeilmail.question.QuestionRepository;
 import maeilmail.question.QuestionSummary;
-import maeilmail.subscribe.core.Subscribe;
-import maeilmail.subscribe.core.SubscribeRepository;
+import maeilmail.subscribe.Subscribe;
+import maeilmail.subscribe.SubscribeRepository;
 import maeilmail.support.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
